backend/requirements.txt
asgiref==3.9.1
Django==4.2.23
django-cors-headers==4.7.0
djangorestframework==3.16.0
numpy==2.0.2
scipy==1.13.1
sqlparse==0.5.3
typing_extensions==4.14.1
